+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * calvin_processor: add full hyperdrive command line to calibration_fits table (calibration_command column).
 * calvin_controller: add s3 config file items: s3_endpoints, s3_profile, s3_bucket.
   * Now uploads any solution images to S3 so they can be displayed by the Django (WS) server. Once uploaded they are deleted.
+* Added standalone util to generate the index.json given a local file path.
 * calvin_processor: clip large gains (c.f. config file item processor->max_gains) before generating plots or writing the solution to the database. 
   * Add the original pre-clipped gains to the 'x_gains_pre_clipped_value ' and 'y_gains_pre_clipped_value' columns of the 'calibration_solutions' table.